add knitr_cache functionality

+++ R/specify_instructions.R
@@ -1,5 +1,5 @@
 #' Specify instructions!
-#' 
+#'
 #' @description This function is not meant to be called directly by the user. It
 #'   is integrated into the function 'execute_instructions'.
 #' @param target_dir_docs For further information and examples run the function
@@ -44,15 +44,15 @@
 #' @param file_ext For further information and examples run the function
 #'   'instructions' and open the file 'df_source_files.rds' within the cache
 #'   directory.
-#' @note This function is part of a family of functions each of which end with 
-#'   '_instructions'. The order to call these functions is: 
-#'   'prepare_instructions', 'implement_instructions', 'check_instructions', 
-#'   'delete_deprecated_instructions', 'execute_instructions' and optionally 
-#'   'output_instructions'. There is a wrapper for these functions called 
+#' @note This function is part of a family of functions each of which end with
+#'   '_instructions'. The order to call these functions is:
+#'   'prepare_instructions', 'implement_instructions', 'check_instructions',
+#'   'delete_deprecated_instructions', 'execute_instructions' and optionally
+#'   'output_instructions'. There is a wrapper for these functions called
 #'   'instructions'.
-#' @seealso \code{\link{prepare_instructions}}, 
-#'   \code{\link{implement_instructions}}, \code{\link{check_instructions}}, 
-#'   \code{\link{instructions}}, \code{\link{execute_instructions}}, 
+#' @seealso \code{\link{prepare_instructions}},
+#'   \code{\link{implement_instructions}}, \code{\link{check_instructions}},
+#'   \code{\link{instructions}}, \code{\link{execute_instructions}},
 #'   \code{\link{output_instructions}}
 #' @author Frederik Sachser
 #' @export
@@ -70,7 +70,8 @@ specify_instructions <-
            figure_source,
            figure_out,
            use_spin,
-           file_ext)
+           file_ext,
+           knitr_cache)
   {
     # deprecated source_files
     files_source_dir <- list.files(source_dir, full.names = TRUE, recursive = TRUE)
@@ -132,7 +133,9 @@ specify_instructions <-
       }
       else {
         # delete deprecated files
-        unlink(x = delete_deprecated_files, recursive = TRUE)
+        if(knitr_cache == FALSE) {
+          unlink(x = delete_deprecated_files, recursive = TRUE)
+        }
 
         # specify instruction "source"
         if (instruction == "source") {
